server/health_system/requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
flask-restful==0.3.9
flask-cors==3.0.10
flask-migrate==3.1.0
flask-jwt-extended==4.5.3
bcrypt==4.0.1
PyJWT==2.3.0
python-dotenv==0.19.0 